+++ examples/draw_bboxes.rs
@@ -51,6 +51,8 @@ fn main() {
         }));
     }
 
-    let img = resvg::render(&rtree, fit_to, None).unwrap();
-    img.save_png(&args[2]).unwrap();
+    let pixmap_size = fit_to.fit_to(rtree.svg_node().size.to_screen_size()).unwrap();
+    let mut pixmap = tiny_skia::Pixmap::new(pixmap_size.width(), pixmap_size.height()).unwrap();
+    resvg::render(&rtree, fit_to, pixmap.as_mut()).unwrap();
+    pixmap.save_png(&args[2]).unwrap();
 }